@@ -1,6 +1,4 @@
-global['CSS'] = null;
-
-const mock = () => {
+const createStorageMock = () => {
   let storage = {};
   return {
     getItem: key => key in storage ? storage[key] : null,
@@ -10,8 +8,9 @@ const mock = () => {
   };
 };
 
-Object.defineProperty(window, 'localStorage', {value: mock()});
-Object.defineProperty(window, 'sessionStorage', {value: mock()});
+Object.defineProperty(window, 'CSS', {value: null});
+Object.defineProperty(window, 'localStorage', {value: createStorageMock()});
+Object.defineProperty(window, 'sessionStorage', {value: createStorageMock()});
 Object.defineProperty(document, 'doctype', {
   value: '<!DOCTYPE html>'
 });
