@@ -1,1 +1,20 @@
-module.exports = require('./setupJest.js');
+'use strict';
+
+require('core-js/es6/reflect');
+require('core-js/es7/reflect');
+require('zone.js/dist/zone.js');
+require('zone.js/dist/proxy.js');
+require('zone.js/dist/sync-test');
+require('zone.js/dist/async-test');
+require('zone.js/dist/fake-async-test');
+require('jest-zone-patch');
+// const AngularSnapshotSerializer = require('./AngularSnapshotSerializer');
+// const HTMLCommentSerializer = require('./HTMLCommentSerializer');
+const getTestBed = require('@angular/core/testing').getTestBed;
+const BrowserDynamicTestingModule = require('@angular/platform-browser-dynamic/testing').BrowserDynamicTestingModule;
+const platformBrowserDynamicTesting = require('@angular/platform-browser-dynamic/testing').platformBrowserDynamicTesting;
+
+getTestBed().initTestEnvironment(
+  BrowserDynamicTestingModule,
+  platformBrowserDynamicTesting()
+);
